refactor : 파일 명명 컨벤션 수정 및 위치 이동 (#34)
