+++ src/panorama/dto/panorama/panorama-request.dto.ts
