+++ src/panorama/dto/panorama/panorama-response.dto.ts
